changing the script to Javascript
